safedoc-webapp/requirements.txt
Flask
pyodbc
azure-cognitiveservices-vision-face
requests
